requirements.txt
Django>=6.0.0
djangorestframework>=3.14.0
pytz>=2023.3
django-cors-headers>=4.0.0

